fix: move Mode to types.ts
