+++ src/app/components/types.ts
@@ -1,0 +1,4 @@
+export enum Mode {
+  Edit,
+  Readonly,
+}
